+++ src/PhotoField/DdPhotoField.jsx
@@ -105,7 +105,7 @@ class PhotoField extends React.Component {
       columns, placeholder, label,
       photoList, required,
       maxUpload, readOnly, className, tip,
-      onImagePreview
+      onImagePreview,
     } = this.props;
     const paneProps = {
       columns,
@@ -119,7 +119,13 @@ class PhotoField extends React.Component {
       tip,
       onPickerClick: () => { this.onPickHandler(); },
       onImageDelete: (index) => { this.handleDeleteImage(index); },
-      onImagePreview: (index) => { onImagePreview ? onImagePreview(index) : this.handlePreview(index); },
+      onImagePreview: (index) => {
+        if (onImagePreview) {
+          onImagePreview(index);
+        } else {
+          this.handlePreview(index);
+        }
+      },
     };
     return (
       <PhotoFieldPane {...paneProps} />