+++ src/SelectField/SelectField.jsx
@@ -73,7 +73,6 @@ class SelectField extends React.Component {
     return (
       <Field
         {...t.props}
-        layout="h"
         icon={icon}
         className={classnames(Context.prefixClass('select-field'), {
           [t.props.className]: !!t.props.className,